- new demo: eLua logo rotation on the LM3S8962 board display :) Quite useless, but I just wanted to show that it can be done. Burn the image and:
# lua /rom/logo.lua

NOTE: I used the Codesourcery compiler for this one. Rumor has it that it is has much faster softfloat operations. If the animation has less than 3 fps, try recompiling with the Codesourcery toolchain.

- new module: bitarray. Fixed array container for bits. Currently only 1, 2 ,4 ,8, 16 and 32 bits are supported, but I'm considering support for arbitrary size bit array at some point. I'll add documentation soon.

And, for the first time, I developed and tested all of this directly under Windows. Once again, just to show that it can be done :)