New build structure experimental branch
This is an EXPERIMENTAL branch on which the work for the new
build structure is done. Don't expect it to have valid code,
currently there are just some experiments that I'm doing in
order to get a feeling of the final implementation. I'm mainly
keeping this as a way to synchronize my work between different
computers. You have been warned :)