Working on LTR consolidation
This changes some data structures in the LTR patch to make
a rotable an actual data type with an identity of its own.
Now a table and a rotable have the same header (CommonHeader)
and one can differentiate between them by using specific
fields in this header alone. The overall functionality remains
the same, but the implementation is much more logical. Plus
this will help with the binary modules support (where the old
implementation wouldn't work).
Also added the global _R as an equivalent of _G for rotables.
It contains all the modules that register as global rotables
at Lua startup.