Added ROM strings
If a Lua file is loaded from ROMFS it can be accessed directly now
(no need to load it to RAM first). This has two distinct effects:

- source files: the source is read directly from ROM
- bytecode files: TString (Lua's string type) was augmented with
  the option to keep the string content (NOT the string value itself,
  just its content) in ROM (no need to read it to RAM first). This
  translates directly to RAM savings.

Still testing this, seems to work fine for now. If OK I'll extend it to
keep certain parts of functions in ROM (notably the bytecode).
Once again, this works only for files in ROMFS (for obvious reasons).