When defining a user character, restore cursor position using getpos()
When defining a user-defined character on the LCD panel, the hardware
forgets the current cursor position on the display.  We used to
restore it by keeping track of the cursor in software, but now that
the LCD firmware has a "report current cursor position" command,
we simply use that to save and restore the cursor position when
defining a char. Code saving: 272 bytes.

The old firmware could not return the current cursor position, so
previously, when defining a user-definable character, it could