Bug fix: Allow more than one call to uart.set_buffer()
Previously, every call except the first one to uart.set_buffer()
with a buffer size > 0 would fail, as platform_cpu_set_int() returns
whether it used to be enabled or not, which is 0 the first time
(which happens to == PLATFORM_INT_OK) but 1 every time after that.
This change makes 0 and 1 success; all failure values are negative
so that people can now change UART buffer sizes.