Use lake as the build system
First attempt as using Steve Donovan's 'lake' as the build system
for eLua. It works, but there are many other things left to do. Will
hold this for now, at least until a newer Lake version gets packaged
as a rock, but this should probably be the solution for future (as
opposed to maintaining our proprietary build system).