New configuration for shell_config
It is now possible to configure each individual command that will be compiled in the shell:

```
shell = { commands = {'ls', 'ver', 'type' } }
```

If not specified, all the available commands will be compiled (which is backwards compatible).

Also, the `advanced_shell`  attribute is gone, replaced with an `advanced` key in the shell
configuration. So write this:

```
shell = { advanced = true }
```

instead of this:

```
advanced_shell = true
```